+++ changelog.md
@@ -4,6 +4,12 @@ New Features
     * Test Signals with ADS1299 using `.testSignal()`
     * Continuous impedance testing, where each sample gets an `impedances` object that is an array of impedances for each
         channel.
+    * OpenBCI Radio Test File
+    * Added Sntp npm module 
+
+Work In Progress
+    * NTP Time Synchronization
     
 Bug fixes
     * Impedance calculations
+    * Readme updates